frontend: Extract DocsViewer from EditorDialog
